requirements.txt
aiohttp>=3.9.0
temporalio>=1.5.0
pymongo>=4.6.0
python-dotenv>=1.0.0
flask-wtf>=1.2.0
email-validator>=2.1.0
aiohttp-jinja2 